fixed error w/ displaying empty sf objects on leaflet - affected LAPD overlay
There seems to be an issue when the sf package is plotting on a leaflet
map a spatial
dataframe where there are no points. Every time I performed a subset of
the LAPD data
where there were no points in the subsetted region, I would get the
message:
"Error: is.matrix(x) || inherits(x, "XY") is not TRUE." After research,
it seems to be
related to this https://github.com/rstudio/leaflet/issues/452 which is
still not resolved
as of 8/29.

To get around this, I created an observe() event paired with an if/else
to check
whether the length of the subset was > 0. If so, it will add Markers to
the plot, and
if not, won't do anything. I set the observeEvent to changes in the
input geography name.